#6 feat: Add file hashing progress updates
Add progress updates
during file hashing operations.

This change introduces a callback to the hashFile method in
crypto-service, which is then used by digest-service,
heal-service, replicate-service and verify-service to update the
display service with the hashing progress, showing a percentage of
completion to the user.

Also, improves log readability by using green color for "POS" logs
and cyan color for "URG" logs.